Raspberry Pi Project (Morse Code Translator)
Our goal in this project was to create a program that would be able to translate the user input from Roman letters to Morse code that would be displayed on the monitor. Our secondary goal was to install a loudspeaker that would play the sound of previously translated user input into morse code. We’ve used built-In libraries, loop, lists, many functions and arrays to complete the project